PyBankService/service/requirements.txt
asyncio
fastapi
requests
starlette
uvicorn
flask[async]
flask
opentelemetry-exporter-otlp
opentelemetry-distro
opentelemetry-distro[otlp]
